ai-projects/projects/clinical-intake-router/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langfuse>=4.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=4.0.0
boto3>=1.34.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
sqlglot>=23.0.0
redis>=5.0.0
